Feature/plis 506 admin panel catalog and users (#28)
feat: implement admin panel catalog and user management (PLIS-506)

- Added admin catalog APIs for products, categories, product media, and home page products.
- Added admin user management APIs with listing, filtering, ban/unban, and bulk operations.
- Added bulk catalog operations for product deletion, price updates, category assignment, and category delete strategies.
- Added storage deletion outbox to make media cleanup reliable after product/media deletion.
- Updated product/order/cart DTOs and pricing helpers to support admin and catalog flows.
- Added Liquibase schema updates and dev integration seed data for admin/catalog flows.
- Added controller, service, integration, and repository-adjacent test coverage for admin catalog/users, storage outbox, pricing, and checkout scenarios.

+++ src/main/java/org/plishka/backend/dto/cart/CartItemSummaryDto.java
@@ -1,14 +1,12 @@
 package org.plishka.backend.dto.cart;
 
-import java.math.BigDecimal;
-
 public record CartItemSummaryDto(
         Long cartItemId,
         Long productId,
         String productName,
         String categoryName,
         Integer quantity,
-        BigDecimal unitPrice,
-        BigDecimal subtotal
+        Long unitPrice,
+        Long subtotal
 ) {
 }
