+++ src/main/java/org/plishka/backend/dto/order/OrderDetailDto.java
@@ -1,6 +1,5 @@
 package org.plishka.backend.dto.order;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -8,7 +7,7 @@ public record OrderDetailDto(
         Long orderId,
         String orderNumber,
         String customerName,
-        BigDecimal totalPrice,
+        Long totalPrice,
         String deliveryCity,
         String phone,
         String notes,
