+++ src/main/java/org/plishka/backend/dto/product/ProductSummaryDto.java
@@ -1,12 +1,10 @@
 package org.plishka.backend.dto.product;
 
-import java.math.BigDecimal;
-
 public record ProductSummaryDto(
         Long productId,
         String name,
         CategoryDto category,
-        BigDecimal price,
+        Long price,
         ProductMediaPreviewDto primaryMedia
 ) {
 }
